remove simple builder from templates, use params by default
