+++ templates/release-0.26/src/main/java/com/badoo/ribs/template/leaf_dagger/foo_bar/builder/FooBarComponent.kt
@@ -4,6 +4,7 @@ import com.badoo.ribs.core.modality.BuildParams
 import com.badoo.ribs.template.leaf_dagger.foo_bar.builder.FooBarModule
 import com.badoo.ribs.template.leaf_dagger.foo_bar.FooBar
 import com.badoo.ribs.template.leaf_dagger.foo_bar.FooBarNode
+import com.badoo.ribs.template.leaf_dagger.foo_bar.builder.FooBarBuilder.Params
 import dagger.BindsInstance
 
 @FooBarScope
@@ -18,7 +19,7 @@ internal interface FooBarComponent {
         fun create(
             dependency: FooBar.Dependency,
             @BindsInstance customisation: FooBar.Customisation,
-            @BindsInstance buildParams: BuildParams<Nothing?>
+            @BindsInstance buildParams: BuildParams<Params>
         ): FooBarComponent
     }
