--- templates/release-0.26/src/main/java/com/badoo/ribs/template/node_dagger_build_param/foo_bar/builder/FooBarScope.kt
@@ -1,7 +1,0 @@
-package com.badoo.ribs.template.node_dagger_build_param.foo_bar.builder
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class FooBarScope
