--- templates/release-0.26/src/main/java/com/badoo/ribs/template/node_dagger_build_param/foo_bar/mapper/InputToWish.kt
@@ -1,10 +1,0 @@
-package com.badoo.ribs.template.node_dagger_build_param.foo_bar.mapper
-
-import com.badoo.ribs.template.node_dagger_build_param.foo_bar.FooBar.Input
-import com.badoo.ribs.template.node_dagger_build_param.foo_bar.feature.FooBarFeature.Wish
-
-internal object InputToWish : (Input) -> Wish? {
-
-    override fun invoke(event: Input): Wish? =
-        null
-}
