--- templates/release-0.26/src/main/java/com/badoo/ribs/template/node_dagger_build_param/foo_bar/mapper/NewsToOutput.kt
@@ -1,10 +1,0 @@
-package com.badoo.ribs.template.node_dagger_build_param.foo_bar.mapper
-
-import com.badoo.ribs.template.node_dagger_build_param.foo_bar.FooBar.Output
-import com.badoo.ribs.template.node_dagger_build_param.foo_bar.feature.FooBarFeature.News
-
-internal object NewsToOutput : (News) -> Output? {
-
-    override fun invoke(news: News): Output? =
-        null
-}
