+++ tooling/rib-intellij-plugin/build.gradle
@@ -53,23 +53,6 @@ templates {
         }
     }
 
-
-    template {
-        id = 'router_node_dagger_build_param_0.26'
-        name = '[0.26] Node + Dagger + BuildParam'
-
-        fromProject = project(':templates:release-0.26')
-        sourcePackage = 'com.badoo.ribs.template.node_dagger_build_param'
-        modulePackage = 'com.badoo.ribs.template'
-        resources = ['rib_foo_bar.xml']
-
-        token {
-            id = "rib_name"
-            name = "RIB name"
-            sourceValue = "FooBar"
-        }
-    }
-
     template {
         id = 'leaf_view_only_release_0.26'
         name = '[0.26] Leaf node (view only)'
